refacto - moving the graphic components from the app file as they are mostly static - a simple utils file containing miscellaneous utilities
